Updated IE, EI, EZ and ZE counter examples
Now all interest points are in the right position, to do a better simulation. For example:

If we are in IE, doing an EI counter example:

In IE we have 105 characters, with AG in position [99,100] (considering list starts in 0), and EI has 12 characters with interest point in [5:6] with GT.

So the counter example will have the interest point in the same position as real data, adding some characters to have the same large.

That means that in position [99,100] will be GT (interest point of EI). But replaced to be a counter example, which is translated in having AG in [99:100]

What we do is take the counter example, expand or reduce it to have the same longitude, put the position interest in the same place as real data, and replace it with the value of actual extractor.